LayoutTests/svg/hittest/svg-embedded-tooltip-expected.txt



Verifying the inline SVG:
svg:
	id:		main_svg
	actual:		'outermost svg'
	expected:	''

use:
	id:		use_without_title
	actual:		'outermost svg'
	expected:	''



Verifying the embedded SVG:
PASS
